nfp: fail probe if serial or interface id is missing
On some platforms with broken ACPI tables we may not have access
to the Serial Number PCIe capability.  This capability is crucial
for us for switchdev operation as we use serial number as switch ID,
and for communication with management FW where interface ID is used.

If we can't determine the Serial Number we have to fail device probe.

+++ src/nfp_plat.c
@@ -305,7 +305,7 @@ static void expa_set(struct nfp_plat *priv, u32 csr, unsigned int id)
 			csr);
 }
 
-static u16 nfp_plat_get_interface(struct device *dev)
+static int nfp_plat_get_interface(struct device *dev)
 {
 	return NFP_CPP_INTERFACE(NFP_CPP_INTERFACE_TYPE_ARM,
 				 0,
